requirements.txt
alembic==1.4.2
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dash==1.10.0
dash-bootstrap-components==0.9.2
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
defusedxml==0.6.0
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Compress==1.4.0
Flask-Dance==3.0.0
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
plotly==4.6.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
URLObject==2.4.3
Werkzeug==1.0.0
WTForms==2.2.1
xlrd==1.2.0
